requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0
ffmpeg-python==0.2.0
python-socketio==5.10.0
aiofiles==23.2.1
asyncio==3.4.3
numpy==1.24.3
python-json-logger==2.0.7
PyYAML==6.0.1 